Rename paper-helper-261-2 to paper-helper-26x and generalize MC 26 version check
