+++ spigot/src/main/java/org/dynmap/bukkit/Helper.java
@@ -40,8 +40,8 @@ public class Helper {
                 Log.info("Loading Glowstone support");
                 BukkitVersionHelper.helper = loadVersionHelper("org.dynmap.bukkit.helper.BukkitVersionHelperGlowstone");
             }
-            else if (v.contains("(MC: 26.1")) {
-	            BukkitVersionHelper.helper = loadVersionHelper("org.dynmap.bukkit.helper.v261_2.BukkitVersionHelperSpigot261_2");
+            else if (v.contains("(MC: 26.")) {
+	            BukkitVersionHelper.helper = loadVersionHelper("org.dynmap.bukkit.helper.v26x.BukkitVersionHelperSpigot26x");
             }
             else if (v.contains("(MC: 1.21)") || v.contains("(MC: 1.21.1)")) {
                 BukkitVersionHelper.helper = loadVersionHelper("org.dynmap.bukkit.helper.v121.BukkitVersionHelperSpigot121");
